Split Activities collection resources
