+++ app/collections/activities/permissions.js
@@ -1,0 +1,35 @@
+import Activities from './';
+
+Activities.allow({
+  insert: function () {
+    return true;
+  },
+  update: function () {
+    // Get current user ID
+    var currentUserId = Meteor.userId();
+
+    // Check if current user has Admin role
+    var currentUserIsAdmin = Roles.userIsInRole(currentUserId, ["admin"]);
+
+    // Only show edit column for users with Admin role
+    if (currentUserIsAdmin) {
+      return true;
+    } else {
+      return false;
+    }
+  },
+  remove: function () {
+    // Get current user ID
+    var currentUserId = Meteor.userId();
+
+    // Check if current user has Admin role
+    var currentUserIsAdmin = Roles.userIsInRole(currentUserId, ["admin"]);
+
+    // Only show edit column for users with Admin role
+    if (currentUserIsAdmin) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+});
